packages/agent-kit/intermediate/brc20-analyst/pyproject.toml
[project]
name = "unisat-brc20-analyst"
version = "0.1.0"
description = "UniSat BRC20 Token Analyst - Professional BRC20 token analysis agent"
authors = [
    {name = "UniSat", email = "dev@unisat.io"}
]
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "agentkit>=0.1.0",
    "veadk>=0.1.0",
    "google-adk>=0.1.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
